ReAdded the forest scene
Readded the scene, if the dash stops working, idk, refresh or something. I don't know why the rigid bodies and camera screw with it beyond it being based on the mouse position relative to the camera. It breaks regularly so ignore it. It's fine. Probably. Just don't change the camera. I dealt with it so it shouldn't need adjustments. If it doesn't work. Idk.