bug fixes again
Changed the camera settings on cave level and cave  rest area, there was a poly collider confiner issue on both and rest had different lens scale. Refitted the damage thing on the bosses second phase, should be in arena now. Changed the reload feature on the game over menu to act similarly to the new game option on the main menu